server/requirements.txt
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.2.1
colorama==0.4.6
contourpy==1.3.0
curl_cffi==0.11.4
cycler==0.12.1
Flask==3.1.1
flask-cors==6.0.1
fonttools==4.54.1
frozendict==2.4.6
html5lib==1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.7
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mplcursors==0.6
mplfinance==0.12.10b0
multitasking==0.0.11
numpy==2.1.3
packaging==24.1
pandas==2.2.3
peewee==3.17.7
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
protobuf==6.31.1
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
scipy==1.14.1
six==1.16.0
soupsieve==2.6
tenacity==9.0.0
tradingview-screener==3.0.0
tzdata==2024.2
urllib3==2.2.3
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
yfinance==0.2.63
gunicorn==23.0.0
python-dotenv==1.1.1
